Add depth recursion to inUseBySubscription and dependsOnSubscriptions
